+++ semgrep/tests/e2e/targets/dependency_aware/ansi.js
@@ -1,0 +1,2 @@
+// We want the rule not to fire, because our code does not match the pattern even though our dependency has a bad version
+const x = 0