+++ semgrep/tests/e2e/test_dependency_aware_rules.py
@@ -16,6 +16,7 @@ import pytest
         ("rules/dependency_aware/ruby-sca.yaml", "dependency_aware/sca.rb"),
         ("rules/dependency_aware/log4shell.yaml", "dependency_aware/log4shell.java"),
         ("rules/dependency_aware/rust-sca.yaml", "dependency_aware/sca.rs"),
+        ("rules/dependency_aware/ansi-html.yaml", "dependency_aware/ansi.js"),
     ],
 )
 def test_dependency_aware_rules(run_semgrep_in_tmp, snapshot, rule, target):