+++ CHANGELOG.md
@@ -20,6 +20,8 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Autofix: Semgrep no longer errors during `--dry-run`s where one fix changes the line numbers in a file that also has a second autofix.
 - Performance regression when running with --debug (#4761)
 - Allow metrics flag and metrics env var at the same time if both are set to the same value (#4703)
+- Scan `yarn.lock` dependencies that do not specify a hash
+- Run `project-depends-on` rules with only `pattern-inside` at their leaves
 
 ## [0.83.0](https://github.com/returntocorp/semgrep/releases/tag/v0.83.0) - 2022-02-24